refactor(github): move mutual followers into own module
